apply format and fix some raw dict checks

+++ qendpoint-core/src/main/java/com/the_qa_company/qendpoint/core/dictionary/impl/section/IntDictionarySectionMap.java
@@ -303,6 +303,6 @@ public class IntDictionarySectionMap implements DictionarySectionPrivate {
 
 	@Override
 	public DictionarySectionType getSectionType() {
-		return DictionarySectionType.FLOAT;
+		return DictionarySectionType.INT;
 	}
 }
